Refactor ha_healthchecker code structure
Now ha_healthchecker is a service. Refactor the code structure
to make the code more readable.

Related-Bug: theopenlab/openlab#218
